Working lambertian diffuse raytracer, some bugs in actual ray tracing that must be fixed and updating image with each sample

+++ BlueLeafRenderer/Renderer/Ray.cpp
@@ -19,6 +19,6 @@ Ray::Ray(Point3 origin, Vector3 direction): origin(origin), direction(direction)
 /// Calculates the t-th point along the direction specified by our vector
 /// using math equation P(t) = A + bt
 /// - Parameter t: The direction value along our ray
-Point3 Ray::at(float t) const {
+Point3 Ray::at(double t) const {
     return origin + direction * t;
 }
