+++ BlueLeafRenderer/Renderer/Ray.hpp
@@ -20,7 +20,7 @@ public:
     Ray();
     Ray(Point3 origin, Vector3 direction);
     
-    Point3 at(float t) const;
+    Point3 at(double t) const;
 };
 
 #endif /* Ray_hpp */
