+++ BlueLeafRenderer/Renderer/Vector3.cpp
@@ -9,23 +9,23 @@
 
 #include <cmath>
 
-/// Default constructor makes zerod float vector
+/// Default constructor makes zerod double vector
 Vector3::Vector3(): x(0), y(0), z(0) {}
 
 
-/// Makes float vector with given x, y, z coordinates
+/// Makes double vector with given x, y, z coordinates
 /// - Parameters:
 ///   - x: x coord
 ///   - y: y coord
 ///   - z: z coord
-Vector3::Vector3(float x, float y, float z): x(x), y(y), z(z) {}
+Vector3::Vector3(double x, double y, double z): x(x), y(y), z(z) {}
 
 /// Length of vector using standard formula
-float Vector3::length() const {
+double Vector3::length() const {
     return std::sqrt(length_squared());
 }
 
 /// Length of vector squared
-float Vector3::length_squared() const {
+double Vector3::length_squared() const {
     return x * x + y * y + z * z;
 }
